rag_server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
ollama>=0.4.0
supabase>=2.10.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
